requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
sqlalchemy>=2.0,<3.0
pymysql>=1.1,<2.0
python-dotenv>=1.0,<2.0
python-jose[cryptography]>=3.3,<4.0
passlib[bcrypt]>=1.7,<2.0
bcrypt>=4.0.1,<4.1
python-multipart>=0.0.9,<1.0
jinja2>=3.1,<4.0